rename sparkopt to sparkoperator
